Added a view controller extension that does all of the UIAlert stuff for conencting.
Added LEDViewContorller which has the UI and all functionallity besides saving colors after app is closed. This is to select the color for an LED

Added a LEDScarf object which will hold all the information about the current LED values on the scarf and where we will edit the LED Values.

Editted the Send Message command in BLEManager such that there is no overflow of commands being sent to the microcontroller. This was done by adding a delay after a command was sent.